Use env extension class instead of creating a helper class
