+++ src/AzureAuth.Test/IEnvExtensionsTest.cs
@@ -6,20 +6,26 @@ namespace AzureAuth.Test
     using FluentAssertions;
 
     using Microsoft.Authentication.AzureAuth;
+    using Microsoft.Authentication.MSALWrapper;
+    using Microsoft.Authentication.TestHelper;
+    using Microsoft.Extensions.Logging;
     using Microsoft.Office.Lasso.Interfaces;
-
+    using Microsoft.Office.Lasso.Telemetry;
     using Moq;
-
+    using NLog.Targets;
     using NUnit.Framework;
 
     public class IEnvExtensionsTest
     {
         private Mock<IEnv> envMock;
+        private ILogger logger;
+        private MemoryTarget logTarget;
 
         [SetUp]
         public void SetUp()
         {
             this.envMock = new Mock<IEnv>();
+            (this.logger, this.logTarget) = MemoryLogger.Create();
         }
 
         [TestCase("1", true)]
@@ -50,6 +56,60 @@ namespace AzureAuth.Test
         {
             this.envMock.Setup(env => env.Get(It.IsAny<string>())).Returns((string)null);
             IEnvExtensions.InteractiveAuthDisabled(this.envMock.Object).Should().BeFalse();
+        }
+
+        [Test]
+        public void ReadAuthModeFromEnvOrSetDefault_ReturnsDefault_WhenEnvVarIsEmpty()
+        {
+            // Arrange
+            envMock.Setup(e => e.Get(It.IsAny<string>())).Returns(string.Empty);
+
+            // Act
+            var result = IEnvExtensions.ReadAuthModeFromEnvOrSetDefault(envMock.Object, new EventData());
+
+            // Assert
+            result.Should().BeEquivalentTo(new[] { AuthMode.Default });
+        }
+
+        [Test]
+        public void ReadAuthModeFromEnvOrSetDefault_ReturnsParsedAuthModes_WhenEnvVarIsValid()
+        {
+            // Arrange
+            envMock.Setup(e => e.Get(It.IsAny<string>())).Returns("Web,Broker");
+
+            // Act
+            var result = IEnvExtensions.ReadAuthModeFromEnvOrSetDefault(envMock.Object, new EventData());
+
+            // Assert
+            result.Should().BeEquivalentTo(new[] { AuthMode.Web, AuthMode.Broker });
+        }
+
+        [Test]
+        public void ReadAuthModeFromEnvOrSetDefault_ReturnsEmpty_WhenEnvVarIsInvalid()
+        {
+            // Arrange
+            envMock.Setup(e => e.Get(It.IsAny<string>())).Returns("InvalidMode");
+
+            // Act
+            var result = IEnvExtensions.ReadAuthModeFromEnvOrSetDefault(envMock.Object, new EventData());
+
+            // Assert
+            result.Should().BeEmpty();
+        }
+
+        [Test]
+        public void ReadAuthModeFromEnvOrSetDefault_AddsEventData_WhenEnvVarIsValid()
+        {
+            // Arrange
+            var eventData = new EventData();
+            envMock.Setup(e => e.Get(It.IsAny<string>())).Returns("Web");
+
+            // Act
+            var result = IEnvExtensions.ReadAuthModeFromEnvOrSetDefault(envMock.Object, eventData);
+
+            // Assert
+            var env_var = $"env_{EnvVars.AuthMode}";
+            eventData.Properties[env_var.ToLower()].Should().Be("Web");
         }
     }
 }
