+++ src/AzureAuth/Commands/CommandAad.cs
@@ -283,9 +283,10 @@ namespace Microsoft.Authentication.AzureAuth.Commands
             }
 
             // If command line options for mode are not specified, then use the environment variables.
-            this.AuthModes ??= AuthModeHelper.ReadAuthModeFromEnvOrSetDefault(env, eventData, logger);
+            this.AuthModes ??= env.ReadAuthModeFromEnvOrSetDefault(eventData);
             if (!this.AuthModes.Any())
             {
+                this.logger.LogError($"Invalid value specified for environment variable {EnvVars.AuthMode}. Allowed values are: {CommandAad.AuthModeHelperText}");
                 return false;
             }
